🔧 Publish a more comprehensive set of build artifacts
Even though the Dockerfile and configuration template are not
technically artifacts resulting from a build, these two files
are necessary in order for downstream deployment tooling to
deploy the service. While these files are also available from
the repository in GitHub, having all required artifacts to
continue deployment available in one place is worth having a
little duplication, and decoupling the downstream services from
having a direct dependency on the source code in GitHub is
a nice benefit.