🎨 Update Dockerfile to use local lndr-server binary if found
This allows for flexible Docker image creation when building locally
or when building from deployment tooling.

If a `lndr-server` binary is found within the project's base directory,
the Docker image will use that. Otherwise, it will download the binary
from the latest successful build of the `master` branch on CircleCI, as
it did before this commit.